Initial link of client and server

+++ server/src/routers/course.py
@@ -6,35 +6,35 @@ router = APIRouter()
 
 
 @router.get("/courses",response_model=list[Course])
-async def list_courses(lang: str, to_lang: str,school_user=Depends(get_school_user)):
+async def list_courses(lang: str | None = None, to_lang: str | None = None, school_user=Depends(get_school_user)):
     query = """SELECT * FROM course.course 
-    WHERE lang = %s AND to_lang = %s 
+    WHERE (%s::varchar IS NULL OR lang = %s)
+    AND (%s::varchar IS NULL OR to_lang = %s)
     AND deleted = false 
-    AND user = %s
-    AND school = %s""" 
-    results = await get_query_results(query, (lang, to_lang, school_user.user_id, school_user.school))
+    AND user_id = %s
+    AND school = %s
+    ORDER BY updated_at DESC""" 
+    results = await get_query_results(query, (lang, lang, to_lang, to_lang, school_user.user_id, school_user.school))
     res = [Course(**row) for row in results]
     return res
 
 @router.get("/course",response_model=Course)
-async def get_course(course_id: int, lang: str, to_lang: str, school_user=Depends(get_school_user)):
+async def get_course(course_id: int, school_user=Depends(get_school_user)):
     query = """SELECT * FROM course.course 
-        WHERE lang = %s AND to_lang = %s 
-        AND deleted = false 
-        AND user = %s
+        WHERE deleted = false 
+        AND user_id = %s
         AND school = %s
         AND course_id = %s
         """ 
-    results = await get_query_results(query, (lang, to_lang, school_user.user_id, school_user.school, course_id))
+    results = await get_query_results(query, (school_user.user_id, school_user.school, course_id))
     if not results:
         raise HTTPException(status_code=404, detail="Course not found")
-    res = [Course(**row) for row in results]
-    return res
+    return Course(**results[0])
 
 @router.post("/course")
 async def create_course(course: Course, school_user=Depends(get_school_user)):
     sql = """
-        INSERT INTO course.course (lang, to_lang, user, school, level)
+        INSERT INTO course.course (lang, to_lang, user_id, school, level)
         VALUES (%s, %s, %s, %s, %s)
         RETURNING * 
         """
@@ -56,7 +56,7 @@ async def update_course( course: Course, school_user=Depends(get_school_user)) -
             title = %s,
             deleted = %s,
         WHERE course_id = %s
-        AND user = %s
+        AND user_id = %s
         AND school = %s
         RETURNING *
     """
@@ -73,7 +73,7 @@ async def delete_course(course: Course, school_user=Depends(get_school_user))->d
     sql = """
         DELETE FROM course.course
         WHERE course_id = %s
-        AND user = %s
+        AND user_id = %s
         AND school = %s
         RETURNING *
     """
